Show percentage in render.js
Calculate the sum of patients in a single LoT: data contains 1 row per graph segment, eg regimen. For instance 30 patients with DVd.
I calculate the sum for each LoT in sums (list of integers) and use the first item  that I call total_patients.
If ShowValues is True, I show 1% if there is less than 1%, otherwise the round value with %.
 Use rawgraph.alluvial_tnxo (to verify) in the chart section of the rawgraph file, export the graph as svg and adjust in https://inkscape.org/release/inkscape-1.3.2/ if necessary